chore: add profiles and programs schemas
