+++ apps/server/src/modules/profiles/index.ts
@@ -1,0 +1,1 @@
+export { router as profilesRouter } from "./profiles.router";
