+++ apps/server/src/modules/programs/index.ts
@@ -1,0 +1,1 @@
+export { router as programsRouter } from "./programs.router";
